Weather App
I have created Weather App project using HTML, CSS , JavaScript . It uses API key to access the weather details from OpenWeatherMap. Weather App contains a user input field for the user, which takes the input of the city name. Once the user enters the city name and clicks on the button, then the API Request is been sent to the OpenWeatherMap and the response is been retrieved in the application which consists of weather, temperature, wind speed, description, etc.